Address final whole-branch review: deps manifest, vacuous test, punctuation loss, EDGAR safety, vendor attribution, and a few cheap hardening fixes
Applies the Critical, Important, and select Minor findings from the final
whole-branch code review of the domain-biased-captioning feature (8-task
branch: Apple Silicon MLX + parakeet-mlx + vendored NeMo CTC word-spotter).

- Add requirements.txt (pinned versions from the working venv) and SETUP.md
  documenting the pyate reference-corpus and setuptools<81 install landmines,
  since no dependency manifest existed anywhere in the repo (Critical).
- Fix tests/test_biased_model.py's tautological vocabulary-stability
  assertion; replace with a real check that the CTC head has
  len(vocabulary) + 1 output classes, matching the blank-index invariant
  the model actually depends on (Important).
- Fix merge.py's merge_spans to carry trailing punctuation tokens (which
  _group_words folds onto the preceding word) into the replacement text
  instead of silently discarding them, which was merging adjacent SRT
  sentences at spotted-term boundaries; add a regression test (Important).
- Also refine merge.py's already-used guard from an index-contiguity check
  to `any(used[first_idx:last_idx+1])`, so it targets clobbering an
  already-claimed word rather than misfiring on zero-duration TDT tokens
  that legitimately have 0% overlap (Minor, same code path as the above).
- Guard eval_dataset.fetch_json against a placeholder SEC_CONTACT and
  cache it with lru_cache to stop re-downloading company_tickers.json on
  every ticker (Important).
- Add Apache-2.0 header to vendor/fscore.py, a vendor/README.md documenting
  upstream source paths and modifications for all three vendored files, and
  a vendor/LICENSE, to satisfy Apache-2.0 attribution requirements on the
  vendored NeMo derivative code (Important).
- Cast to float32 before the CTC logits -> numpy conversion in
  biased_model.py, removing a latent dependency on log_softmax's upstream
  upcast behavior (Minor).
- Make term_extraction.extract_terms's top_n default actually read from
  config.TERM_EXTRACTION_TOP_N instead of duplicating the literal (Minor).
- Pass chunk_duration=120.0 in eval.py's transcribe_clips, matching the
  CLI's chunking so eval exercises the same code path (Minor).

+++ captionlm/biased_model.py
@@ -45,7 +45,7 @@ class BiasedParakeetTDTCTC(ParakeetTDTCTC):
 
             hypotheses = []
             for batch_idx, tokens in enumerate(result):
-                logprobs = np.array(ctc_logits[batch_idx])
+                logprobs = np.array(ctc_logits[batch_idx].astype(mx.float32))
                 ws_hyps = run_word_spotter(
                     logprobs,
                     self.context_graph,
